Try to fix character create issues.

+++ src/game/Handlers/CharacterHandler.cpp
@@ -304,11 +304,20 @@ void WorldSession::HandleCharCreateOpcode(WorldPacket & recv_data)
         return;
     }
 
-    if (sObjectMgr.GetPlayerGuidByName(name))
+    if (ObjectGuid existingGuid = sObjectMgr.GetPlayerGuidByName(name))
     {
-        data << (uint8)CHAR_CREATE_NAME_IN_USE;
-        SendPacket(&data);
-        return;
+        PlayerCacheData const* pExistingData = sObjectMgr.GetPlayerDataByGUID(existingGuid.GetCounter());
+        if (pExistingData && pExistingData->sName == name)
+        {
+            data << (uint8)CHAR_CREATE_NAME_IN_USE;
+            SendPacket(&data);
+            return;
+        }
+        else
+        {
+            sObjectMgr.DeletePlayerNameFromCache(name);
+            sLog.outError("Character name %s taken but no player data in cache!", name.c_str());
+        }
     }
 
     if (_charactersCount >= sWorld.getConfig(CONFIG_UINT32_CHARACTERS_PER_REALM))
@@ -359,31 +368,34 @@ void WorldSession::HandleCharCreateOpcode(WorldPacket & recv_data)
         }
     }
 
-    Player *pNewChar = new Player(this);
+    // created only to call SaveToDB()
+    std::unique_ptr<Player> pNewChar = std::make_unique<Player>(this);
     if (!pNewChar->Create(sObjectMgr.GeneratePlayerLowGuid(), name, race_, class_, gender, skin, face, hairStyle, hairColor, facialHair))
     {
         // Player not create (race/class problem?)
-        delete pNewChar;
-
         data << (uint8)CHAR_CREATE_ERROR;
         SendPacket(&data);
-
         return;
     }
 
     MasterPlayer masterPlayer(this);
-    masterPlayer.Create(pNewChar);
+    masterPlayer.Create(pNewChar.get());
     if ((have_same_race && skipCinematics == CINEMATICS_SKIP_SAME_RACE) || skipCinematics == CINEMATICS_SKIP_ALL)
         pNewChar->SetCinematic(1);                          // not show intro
 
     pNewChar->SetAtLoginFlag(AT_LOGIN_FIRST);               // First login
 
     // Player created, save it now
-    pNewChar->SaveToDB();
+    if (!pNewChar->SaveToDB(false, true, true))
+    {
+        data << (uint8)CHAR_CREATE_ERROR;
+        SendPacket(&data);
+        return;
+    }
     masterPlayer.SaveToDB();
 
-    sObjectMgr.InsertPlayerInCache(pNewChar);
-    sObjectMgr.UpdatePlayerCachedPosition(pNewChar);
+    sObjectMgr.InsertPlayerInCache(pNewChar.get());
+    sObjectMgr.UpdatePlayerCachedPosition(pNewChar.get());
     _charactersCount += 1;
 
     LoginDatabase.PExecute("DELETE FROM realmcharacters WHERE acctid= '%u' AND realmid = '%u'", GetAccountId(), realmID);
@@ -396,7 +408,6 @@ void WorldSession::HandleCharCreateOpcode(WorldPacket & recv_data)
     BASIC_LOG("Account: %d (IP: %s) Create Character:[%s] (guid: %u)", GetAccountId(), IP_str.c_str(), name.c_str(), pNewChar->GetGUIDLow());
     sLog.out(LOG_CHAR, "[%s:%u@%s] Create Character:[%s] (guid: %u)", GetUsername().c_str(), GetAccountId(), IP_str.c_str(), name.c_str(), pNewChar->GetGUIDLow());
     sDBLogger->LogCharAction({ pNewChar->GetGUIDLow(), GetAccountId(), LogCharAction::ActionCreate, {} });
-    delete pNewChar;                                        // created only to call SaveToDB()
 }
 
 void WorldSession::HandleCharDeleteOpcode(WorldPacket & recv_data)
